Add support for SCSS
Although there was a support for SASS files already present, it did not work in case when file had a .SCSS extension.

This was due to the fact that JetBrains has a dedicated implementation for SCSS, so `SASSLanguage` was not recognized as a proper `FocusedElementFinder`.

This commit fixes it by introducing additional `ScssFocusedElementFinder` which mimics the behaviouir of `SassFocusedElementFinder`.